pyproject.toml
[tool.poetry]
name = "delta-lake-library"
version = "0.1.0"
description = "Delta lake library"
authors = ["Tam Nguyen <npt.dc@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.11"
types-requests = "^2.27.7"
types-python-dateutil = "^2.8.7"
databricks-cli = "^0.16.2"
databricks-connect = "^9.1.5"
databricks-sql-connector = "^0.9.4"
confluent-kafka = "^1.8.2"
python-dotenv = "^0.19.2"
numpy = "^1.22.1"
pandas = "^1.3.5"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
mypy = "^0.931"
flake8 = "^4.0.1"
pep8-naming = "^0.12.1"
pydocstyle = "^6.1.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pre-commit = "^2.16.0"
jupyterlab = "^3.2.8"
jupyterthemes = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
skip_glob = ["**/notebooks/*", "**/venv/*"]

[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = "notebooks|venv"

[tool.pydocstyle]
ignore = ["D104", "D107", "D212", "D203"]
match-dir = "^(?!notebooks|venv).*"
